requirements.txt
python-dateutil==2.4.2
gunicorn==19.3.0
simplejson==3.7.3
unittest2==0.5.1
Flask>=0.9
Werkzeug==0.8.3
Jinja2==2.6
webassets==0.7
Fabric==1.4.3
ssh==1.7.14
pycrypto==2.6
paramiko==1.15.2
sh==1.11
pytest
twitter==1.17.0
pytz==2015.4
Flask-Testing==0.4.2
facebook-sdk
markdown
premailer
requests
retrying